backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyflakes==3.1.0
pylint==3.0.3
pytest==7.4.3
httpx==0.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
alembic==1.12.1

